# test.github.io-master/server/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 云上共富 - 农村集体经济数字化平台 API
  description: |
    云上共富购物系统后端API文档

    ## 核心功能模块
    - 用户认证模块 (/api/auth)
    - 商品管理模块 (/api/products)
    - 订单模块 (/api/orders)
    - 投资项目模块 (/api/projects)
    - 投标模块 (/api/bids)
    - 用户管理模块 (/api/users)
  version: 1.0.0
  contact:
    name: 云上共富技术支持
servers:
  - url: http://localhost:3001
    description: 开发环境服务器
...
paths:
...
  /api/users:
    get:
      tags:
        - 用户
      summary: 获取用户列表（仅管理员）
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
  /api/users/{id}:
    get:
      tags:
        - 用户
      summary: 获取用户详情
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...